+++ changelog.md
@@ -1,5 +1,11 @@
 # Change Log for GuestConfiguration module
 
+## v1.20.0
+
+- Ensures shell script to install Inspec has Linux line endings
+- Include PowerShell modules when they are required for DSC modules in packages
+- More target file copy when creating custom packages
+
 ## v1.19.3
 
 - Bug fix for New-GuestConfigurationPolicy not generating correct Linux audit IF